requirements.txt
streamlit
pymongo[srv]
plotly
pandas
certifi
python-dotenv
requests
pydantic
networkx